refactor(chat.css): delete unused class names
