+++ src/Components/ChatComponent/chat.css
@@ -1,177 +1,16 @@
 .chat-wrapper {
-  display: flex;
   flex: 1;
-  position: relative;
-  height: 100%;
-}
-
-.chat-area {
-  border-left: 1px solid #ccc;
-  border-right: 1px solid #ccc;
-}
-
-.chat-messages {
-  overflow-y: auto;
-  scroll-behavior: smooth;
 }
 
 .chat-container {
+  /* TODO: Use `overflow: hidden` as a temporary
+      workaround, further investigation is needed
+      to identify the root cause.
+  */
   display: flex;
   height: 100%;
   width: 100%;
   overflow: hidden;
-}
-
-.manager-send-message-container .task-select {
-  border: 1px solid #ddd;
-}
-
-.manager-send-message-container .message-options label {
-  color: black;
-}
-
-.manager-send-message-container {
-  display: flex;
-  padding: 10px;
-  background-color: #fff;
-  border-top: 1px solid #ddd;
-  gap: 10px;
-  height: auto;
-  align-items: flex-start;
-}
-
-.text-area-message {
-  flex: 7;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-  resize: none;
-  outline: none;
-  min-height: 138px;
-  width: 100%;
-  font-size: 18px;
-}
-
-.manager-send-message-container .input-group {
-  gap: 0;
-}
-
-.message-options {
-  flex: 3;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 10px;
-}
-
-.button-row {
-  display: flex;
-  justify-content: space-between;
-  gap: 5px;
-  flex: 1;
-}
-
-.action-button {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  border: 1px solid #ddd;
-  padding: 15px;
-  border-radius: 5px;
-  transition: background-color 0.3s;
-  font-size: 16px;
-  cursor: pointer;
-}
-
-.action-button:hover {
-  background-color: #0f824c;
-  color: #fff;
-}
-
-.select-row {
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-  flex: 1;
-}
-
-.task-select {
-  width: 100%;
-}
-
-.task-select:focus {
-  border-color: #0f824c;
-  box-shadow: 0 0 5px rgba(15, 130, 76, 0.5);
-}
-
-.container-options-component {
-  text-align: left;
-}
-
-.select-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  gap: 10px;
-  align-items: center;
-}
-
-.select-container label {
-  margin-right: 10px;
-  font-weight: bold;
-  color: #555;
-}
-
-.empty-chat {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  width: 100%;
-  font-size: 24px;
-  letter-spacing: 1px;
-}
-
-.message-content {
-  display: flex;
-  flex-direction: column;
-}
-
-.message-row {
-  display: flex;
-  align-items: center;
-}
-
-.reaction-toggle-button {
-  display: flex;
-  cursor: pointer;
-  align-items: center;
-  justify-content: flex-end;
-  font-size: 18px;
-}
-
-.reaction-buttons {
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-}
-
-.reaction-buttons div {
-  cursor: pointer;
-  font-size: 18px;
-}
-
-.text img {
-  cursor: pointer;
-  width: auto;
-  background-color: white;
-  color: black;
-}
-
-.tabs {
-  display: flex;
 }
 
 .text-message {
