+++ src/Components/ChatComponent/components/ChatMessages/index.js
@@ -1,0 +1,1 @@
+export * from "./ChatMessages"
