+++ src/Components/ChatComponent/components/index.js
@@ -5,3 +5,4 @@ export * from "./PersonalData4ClientForm"
 export * from "./ChatInput"
 export * from "./GroupedMessages"
 export * from "./EmojiMessage"
+export * from "./ChatMessages"
